lex: connect to the build
Another case of using integers to store pointers (or integers).  Eugh.

I have tested this, in an almost valid fashion, by building a 64bit x86 lex
with these fixes, using it to build illumos, and then comparing both the
results and all the sources generated by the build (any source file `git
status -u` lists), both were clean.  I have also compared libsec's ACL lexer
generated on arm v. x86.

+++ usr/src/cmd/sgs/lex/common/once.h
@@ -106,8 +106,8 @@ int chset;	/* 1 = char set modified */
 FILE *fin, *fother;
 int fptr;
 int *name;
-int *left;
-int *right;
+intptr_t *left;
+intptr_t *right;
 int *parent;
 Boolean *nullstr;
 int tptr;
